[CCUBE-2148][ZZ] refactor: convert forcing theme to separate hook
