+++ src/e-signature/e-signature-canvas.tsx
@@ -6,11 +6,10 @@ import {
     useEffect,
     useImperativeHandle,
     useRef,
-    useState,
 } from "react";
 
-import { Colour, useTheme } from "../theme";
-import { parseCSSVariableValue } from "../theme/utils/css-variable";
+import { Colour } from "../theme";
+import { useForcedModeToken } from "../util";
 import * as styles from "./e-signature.styles";
 
 interface ESignatureCanvasProps {
@@ -35,23 +34,7 @@ const Component = (
     const canvasRef = useRef<HTMLCanvasElement>(null);
     const fabricCanvas = useRef<FabricCanvas>();
     const pencilBrush = useRef<PencilBrush>();
-    const { theme, themeElement } = useTheme();
-    const [resolvedColor, setResolvedColor] = useState<string | undefined>();
-
-    useEffect(() => {
-        if (!themeElement) return;
-
-        const tempEl = document.createElement("div");
-        tempEl.setAttribute("data-fds-theme", theme);
-        tempEl.setAttribute("data-fds-theme-mode", "light");
-        tempEl.style.display = "none";
-        themeElement.appendChild(tempEl);
-
-        const value = parseCSSVariableValue(Colour["text"], tempEl);
-        setResolvedColor(value || undefined);
-
-        themeElement.removeChild(tempEl);
-    }, [theme, themeElement, resolvedColor]);
+    const resolvedColor = useForcedModeToken(Colour["text"]);
 
     // =============================================================================
     // HOOKS
